// src/types/nonsig.d.ts
export namespace NonsigTypes {
  export interface ICustomerInfo {
    nsTradeStyle?: string
    nonsig?: string
    nsAddr1?: string
    nsAddr2?: string
    nsCity?: string
    nsState?: string
    nsPostalCode?: string
    nsCountry?: string
    nsBrandKey?: string
    active?: boolean
    active_thq?: boolean
    nsType?: string
  }
}